int16/int32

+++ tests/test_backgammon.py
@@ -139,9 +139,9 @@ def test_is_turn_end():
         rng=rng,
         board=board,
         turn=jnp.int32(1),
-        dice=jnp.array([2, 2], dtype=jnp.int16),
-        playable_dice=jnp.array([-1, -1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([2, 2], dtype=jnp.int32),
+        playable_dice=jnp.array([-1, -1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
     )
     assert _is_turn_end(state)
 
@@ -152,9 +152,9 @@ def test_is_turn_end():
         rng=rng,
         board=board,
         turn=jnp.int32(1),
-        dice=jnp.array([2, 2], dtype=jnp.int16),
-        playable_dice=jnp.array([-1, -1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(2),
+        dice=jnp.array([2, 2], dtype=jnp.int32),
+        playable_dice=jnp.array([-1, -1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(2),
     )
     assert _is_turn_end(state)
 
@@ -181,9 +181,9 @@ def test_change_turn():
         rng=rng,
         board=test_board,
         turn=jnp.int32(0),
-        dice=jnp.array([2, 2], dtype=jnp.int16),
-        playable_dice=jnp.array([-1, -1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(2),
+        dice=jnp.array([2, 2], dtype=jnp.int32),
+        playable_dice=jnp.array([-1, -1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(2),
     )
     state = _change_turn(state)
     print(state._board, board)
@@ -193,16 +193,16 @@ def test_change_turn():
 def test_no_op():
     board: jnp.ndarray = make_test_boad()
     legal_action_mask = _legal_action_mask(
-        board, jnp.array([0, 1, -1, -1], dtype=jnp.int16)
+        board, jnp.array([0, 1, -1, -1], dtype=jnp.int32)
     )
     state = make_test_state(
         current_player=jnp.int32(1),
         rng=rng,
         board=board,
         turn=jnp.int32(1),
-        dice=jnp.array([0, 1], dtype=jnp.int16),
-        playable_dice=jnp.array([0, 1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([0, 1], dtype=jnp.int32),
+        playable_dice=jnp.array([0, 1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
         legal_action_mask=legal_action_mask,
     )
     state = step(state, 0)  # execute no-op action
@@ -214,16 +214,16 @@ def test_step():
     board: jnp.ndarray = make_test_boad()
     board = _flip_board(board)  # Flipped
     legal_action_mask = _legal_action_mask(
-        board, jnp.array([0, 1, -1, -1], dtype=jnp.int16)
+        board, jnp.array([0, 1, -1, -1], dtype=jnp.int32)
     )
     state = make_test_state(
         current_player=jnp.int32(1),
         rng=rng,
         board=board,
         turn=jnp.int32(1),
-        dice=jnp.array([0, 1], dtype=jnp.int16),
-        playable_dice=jnp.array([0, 1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([0, 1], dtype=jnp.int32),
+        playable_dice=jnp.array([0, 1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
         legal_action_mask=legal_action_mask,
     )
     expected_legal_action_mask: jnp.ndarray = jnp.zeros(
@@ -244,7 +244,7 @@ def test_step():
     # White plays die=2 24(bar)->1
     state = step(state=state, action=(1) * 6 + 1)
     assert (
-            state._playable_dice == jnp.array([0, -1, -1, -1], dtype=jnp.int16)
+            state._playable_dice == jnp.array([0, -1, -1, -1], dtype=jnp.int32)
     ).all()  # Is playable dice updated correctly?
     assert state._played_dice_num == 1  # played dice increased?
     assert state._turn == 1  # turn is not changed?
@@ -267,16 +267,16 @@ def test_step():
     # black
     board: jnp.ndarray = make_test_boad()
     legal_action_mask = _legal_action_mask(
-        board, jnp.array([4, 5, -1, -1], dtype=jnp.int16)
+        board, jnp.array([4, 5, -1, -1], dtype=jnp.int32)
     )
     state = make_test_state(
         current_player=jnp.int32(0),
         rng=rng,
         board=board,
         turn=jnp.int32(0),
-        dice=jnp.array([4, 5], dtype=jnp.int16),
-        playable_dice=jnp.array([4, 5, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([4, 5], dtype=jnp.int32),
+        playable_dice=jnp.array([4, 5, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
         legal_action_mask=legal_action_mask,
     )
     expected_legal_action_mask: jnp.ndarray = jnp.zeros(
@@ -305,9 +305,9 @@ def test_observe():
         rng=rng,
         board=board,
         turn=jnp.int32(1),
-        dice=jnp.array([0, 1], dtype=jnp.int16),
-        playable_dice=jnp.array([0, 1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([0, 1], dtype=jnp.int32),
+        playable_dice=jnp.array([0, 1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
     )
     expected_obs = jnp.concatenate(
         (board, jnp.array([1, 1, 0, 0, 0, 0])), axis=None
@@ -319,9 +319,9 @@ def test_observe():
         rng=rng,
         board=board,
         turn=jnp.int32(1),
-        dice=jnp.array([0, 1], dtype=jnp.int16),
-        playable_dice=jnp.array([1, 1, 1, 1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([0, 1], dtype=jnp.int32),
+        playable_dice=jnp.array([1, 1, 1, 1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
     )
     expected_obs = jnp.concatenate(
         (board, jnp.array([0, 4, 0, 0, 0, 0])), axis=None
@@ -334,9 +334,9 @@ def test_observe():
         rng=rng,
         board=board,
         turn=jnp.int32(-1),
-        dice=jnp.array([0, 1], dtype=jnp.int16),
-        playable_dice=jnp.array([-1, 1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([0, 1], dtype=jnp.int32),
+        playable_dice=jnp.array([-1, 1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
     )
     expected_obs = jnp.concatenate(
         (board, jnp.array([0, 1, 0, 0, 0, 0])), axis=None
@@ -348,9 +348,9 @@ def test_observe():
         rng=rng,
         board=board,
         turn=jnp.int32(-1),
-        dice=jnp.array([0, 1], dtype=jnp.int16),
-        playable_dice=jnp.array([-1, 1, -1, -1], dtype=jnp.int16),
-        played_dice_num=jnp.int16(0),
+        dice=jnp.array([0, 1], dtype=jnp.int32),
+        playable_dice=jnp.array([-1, 1, -1, -1], dtype=jnp.int32),
+        played_dice_num=jnp.int32(0),
     )
     expected_obs = jnp.concatenate(
         (1 * board, jnp.array([0, 0, 0, 0, 0, 0])), axis=None
@@ -492,7 +492,7 @@ def test_move():
 def test_legal_action():
     board = make_test_boad()
     # black
-    playable_dice = jnp.array([3, 2, -1, -1], dtype=jnp.int16)
+    playable_dice = jnp.array([3, 2, -1, -1], dtype=jnp.int32)
     expected_legal_action_mask: jnp.ndarray = jnp.zeros(
         6 * 26, dtype=jnp.bool_
     )
@@ -521,7 +521,7 @@ def test_legal_action():
     print(jnp.where(expected_legal_action_mask != 0)[0])
     assert (expected_legal_action_mask == legal_action_mask).all()
 
-    playable_dice = jnp.array([5, 5, 5, 5], dtype=jnp.int16)
+    playable_dice = jnp.array([5, 5, 5, 5], dtype=jnp.int32)
     expected_legal_action_mask = jnp.zeros(6 * 26, dtype=jnp.bool_)
     expected_legal_action_mask = expected_legal_action_mask.at[
         6 * (19 + 2) + 5
@@ -531,7 +531,7 @@ def test_legal_action():
 
     # white
     board = _flip_board(board)
-    playable_dice = jnp.array([4, 1, -1, -1], dtype=jnp.int16)
+    playable_dice = jnp.array([4, 1, -1, -1], dtype=jnp.int32)
     expected_legal_action_mask: jnp.ndarray = jnp.zeros(
         6 * 26, dtype=jnp.bool_
     )
@@ -541,7 +541,7 @@ def test_legal_action():
     legal_action_mask = _legal_action_mask(board, playable_dice)
     assert (expected_legal_action_mask == legal_action_mask).all()
 
-    playable_dice = jnp.array([4, 4, 4, 4], dtype=jnp.int16)
+    playable_dice = jnp.array([4, 4, 4, 4], dtype=jnp.int32)
     expected_legal_action_mask = jnp.zeros(
         6 * 26, dtype=jnp.bool_
     )  # dance
@@ -554,20 +554,20 @@ def test_legal_action():
 
 def test_calc_win_score():
     # backgammon win by black
-    back_gammon_board = jnp.zeros(28, dtype=jnp.int16)
+    back_gammon_board = jnp.zeros(28, dtype=jnp.int32)
     back_gammon_board = back_gammon_board.at[26].set(15)
     back_gammon_board = back_gammon_board.at[23].set(-15)  # black on home board
     print(_calc_win_score(back_gammon_board))
     assert _calc_win_score(back_gammon_board) == 3
 
     # gammon win by black
-    gammon_board = jnp.zeros(28, dtype=jnp.int16)
+    gammon_board = jnp.zeros(28, dtype=jnp.int32)
     gammon_board = gammon_board.at[26].set(15)
     gammon_board = gammon_board.at[7].set(-15)
     assert _calc_win_score(gammon_board) == 2
 
     # single win by black
-    single_board = jnp.zeros(28, dtype=jnp.int16)
+    single_board = jnp.zeros(28, dtype=jnp.int32)
     single_board = single_board.at[26].set(15)
     single_board = single_board.at[27].set(-3)
     single_board = single_board.at[3].set(-12)
@@ -575,7 +575,7 @@ def test_calc_win_score():
 
 
 def test_black_off():
-    board: jnp.ndarray = jnp.zeros(28, dtype=jnp.int16)
+    board: jnp.ndarray = jnp.zeros(28, dtype=jnp.int32)
     board = board.at[0].set(15)
     playable_dice = jnp.array([3, 2, -1, -1])
     legal_action_mask = _legal_action_mask(board, playable_dice)
